# [CHIA-2052] TreeCache optimization