Pairs don't need to be hashed, since all atoms are deduplicated identical pairs have the same child nodes, so they can just be deduplicated based on those. only atoms are hashed for deduplication.

This improves timing with up to 25%, but more commonly in the 10% range in the benchmarks. 

MacBook M1:
```
group            optimized-tree-cache    tree-cache
-----            --------------------    ----------
Serializer 0     1.00    131.8±7.03ms    1.18    156.1±6.97ms
Serializer 1     1.00    112.8±2.90ms    1.13    127.3±3.73ms
Serializer 2     1.00    109.6±0.92µs    1.33    145.3±1.23µs
Serializer 3     1.00  1870.4±64.51µs    1.09      2.0±0.01ms
Serializer 4     1.00    235.1±0.96µs    1.11    260.2±2.36µs
```

RPi-5:
```
group            optimized-tree-cache    main
-----            --------------------    ----------
Serializer 0     1.00   453.1±41.70ms    1.14   518.4±49.44ms
Serializer 1     1.00   366.0±32.49ms    1.13   414.5±32.71ms
Serializer 2     1.00   362.0±59.69µs    1.23   444.9±68.23µs
Serializer 3     1.00      5.2±0.91ms    1.10      5.7±0.95ms
Serializer 4     1.00  750.2±113.10µs    1.02  761.9±118.06µs
```